Moved DBPinnableSlice to dedicated module...
