+++ src/db.rs
@@ -27,8 +27,6 @@ use libc::{self, c_char, c_void, size_t};
 use std::collections::BTreeMap;
 use std::ffi::{CStr, CString};
 use std::fmt;
-use std::marker::PhantomData;
-use std::ops::Deref;
 use std::path::{Path, PathBuf};
 use std::slice;
 use std::str;
@@ -731,56 +729,6 @@ impl Drop for DB {
 impl fmt::Debug for DB {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "RocksDB {{ path: {:?} }}", self.path())
-    }
-}
-
-/// Wrapper around RocksDB PinnableSlice struct.
-///
-/// With a pinnable slice, we can directly leverage in-memory data within
-/// RocksDB toa void unnecessary memory copies. The struct here wraps the
-/// returned raw pointer and ensures proper finalization work.
-pub struct DBPinnableSlice<'a> {
-    ptr: *mut ffi::rocksdb_pinnableslice_t,
-    db: PhantomData<&'a DB>,
-}
-
-impl<'a> AsRef<[u8]> for DBPinnableSlice<'a> {
-    fn as_ref(&self) -> &[u8] {
-        // Implement this via Deref so as not to repeat ourselves
-        &*self
-    }
-}
-
-impl<'a> Deref for DBPinnableSlice<'a> {
-    type Target = [u8];
-
-    fn deref(&self) -> &[u8] {
-        unsafe {
-            let mut val_len: size_t = 0;
-            let val = ffi::rocksdb_pinnableslice_value(self.ptr, &mut val_len) as *mut u8;
-            slice::from_raw_parts(val, val_len)
-        }
-    }
-}
-
-impl<'a> Drop for DBPinnableSlice<'a> {
-    fn drop(&mut self) {
-        unsafe {
-            ffi::rocksdb_pinnableslice_destroy(self.ptr);
-        }
-    }
-}
-
-impl<'a> DBPinnableSlice<'a> {
-    /// Used to wrap a PinnableSlice from rocksdb to avoid unnecessary memcpy
-    ///
-    /// # Unsafe
-    /// Requires that the pointer must be generated by rocksdb_get_pinned
-    pub unsafe fn from_c(ptr: *mut ffi::rocksdb_pinnableslice_t) -> DBPinnableSlice<'a> {
-        DBPinnableSlice {
-            ptr,
-            db: PhantomData,
-        }
     }
 }
